CoRaLS Documentation

Welcome to the CoRaLS documentation

CoRaLS — Module

CoRaLS

The CoRaLS (Cosmic Ray Lunar Sounder) module is designed for computing acceptance rates from cosmic ray interactions in lunar regolith.

Installation

Install the latest version of Julia from julialang.org/downloads and add Julia to your system path.

Also ensure python and the matplotlib package are installed with Anaconda/Miniconda/Miniforge (e.g. so that it is visible with conda list matplotlib).

Clone this repository:

git clone git@github.com:cjtu/CoRaLS.jl.git

Change directory to the cloned repository and start Julia.

cd CoRaLS.jl
julia

In the Julia REPL, enter package mode by pressing ] (notice julia> changes to >pkg). Adding and building the PyCall package will give Julia access to the Python installation specified at the path supplied in quotes to ENV["PYTHON"] (leave blank "" to use system default python):

julia> ENV["PYTHON"]=""
julia> ]
pkg> add PyCall
pkg> build PyCall

Then activate the CoRaLS environment and instantiate it to install the required packages and compile CoRaLS.jl. You will need to activate the environment each time you start a new Julia session:

pkg> activate .
pkg> instantiate

Now CoRaLS is ready to use. To exit package mode, press backspace (prompt is now julia>). Try importing CoRaLS and running a test:

julia> using CoRaLS
julia> plot_acceptance(acceptance(10000, 20))

Or equivalently:

julia> import CoRaLS
julia> CoRaLS.plot_acceptance(CoRaLS.acceptance(10000, 20))

Exit julia with ctrl+D or exit():

julia> exit()

Using the CoRaLS environment

From now on, simply start julia with the environment activated with --project=/path/to/root/CoRaLS/directory. If you are in the folder simply run:

julia --project=.

Then you can import CoRaLS and use it as above:

julia> using CoRaLS

Testing

To test the installation, run the following from the CoRaLS directory:

julia --project=. test/runtests.jl

Developers

To build documentation, run:

julia --project=docs/ docs/make.jl

To preview docs locally, run, then go to the url in a browser:

julia --project=docs/ -e 'using LiveServer; serve(dir="docs/build")'

Citing

Paper coming soon.

API Index

CoRaLS

CoRaLS.Rmoon

CoRaLS.ARW

CoRaLS.AbstractSignal

CoRaLS.Acceptance

CoRaLS.ConstantDensity

CoRaLS.ConstantIndex

CoRaLS.Direct

CoRaLS.DivergenceModel

CoRaLS.EventGeometry

CoRaLS.FORTE

CoRaLS.FarFieldDivergence

CoRaLS.FieldModel

CoRaLS.GaisserHillasProfile

CoRaLS.GaussianIceRoughness

CoRaLS.GaussianProfile

CoRaLS.GaussianRoughness

CoRaLS.GaussianSlope

CoRaLS.GeometryImplementation

CoRaLS.IceRoughnessModel

CoRaLS.JAM

CoRaLS.LunarPole

CoRaLS.MixedFieldDivergence

CoRaLS.NearFieldDivergence

CoRaLS.NoIceRoughness

CoRaLS.NoRoughness

CoRaLS.NoSlope

CoRaLS.OldIncorrectDensity

CoRaLS.ProfileModel

CoRaLS.Reflected

CoRaLS.RegolithDensity

CoRaLS.RegolithIndex

CoRaLS.RoughnessModel

CoRaLS.ScalarGeometry

CoRaLS.SlopeModel

CoRaLS.StrangwayDensity

CoRaLS.StrangwayIndex

CoRaLS.SurfaceDeepIndex

CoRaLS.TrialFailed

CoRaLS.VectorGeometry

Base.show

CoRaLS.ANITA

CoRaLS.ARW_RE

CoRaLS.LPDA

CoRaLS.acceptance

CoRaLS.attenuation_length

CoRaLS.auger_spectrum

CoRaLS.auger_spectrum

CoRaLS.auger_spectrum_2020

CoRaLS.auger_spectrum_2021

CoRaLS.cartesian_to_latlon

CoRaLS.cartesian_to_latlonalt

CoRaLS.cartesian_to_spherical

CoRaLS.charge_excess_profile

CoRaLS.compute_direct

CoRaLS.compute_reflected

CoRaLS.create_antenna

CoRaLS.create_density_lut

CoRaLS.differential_spectrum

CoRaLS.divergence_tpar

CoRaLS.divergence_tperp

CoRaLS.estimate_Xmax

CoRaLS.extragalactic_noise

CoRaLS.fresnel_coeffs

CoRaLS.fresnel_critical

CoRaLS.galactic_noise

CoRaLS.gaussian_trigger

CoRaLS.horizon_angle

CoRaLS.ice_roughness

CoRaLS.intersect_with_sphere

CoRaLS.latlon_to_cartesian

CoRaLS.latlon_to_xy

CoRaLS.load_ARW

CoRaLS.load_acceptance

CoRaLS.magnitude_trigger

CoRaLS.merge_acceptance

CoRaLS.old_acceptance

CoRaLS.parse_orbit

CoRaLS.plot_acceptance

CoRaLS.plot_differential_spectrum

CoRaLS.plot_incident_angles

CoRaLS.plot_offaxis_angle

CoRaLS.plot_polarization_angle

CoRaLS.point_impacts_psr

CoRaLS.polarization_angle

CoRaLS.propagate_and_refract

CoRaLS.propagate_cosmicray

CoRaLS.propagate_to_Xmax

CoRaLS.psr_area

CoRaLS.psr_fraction

CoRaLS.random_angles_on_cap

CoRaLS.random_direction

CoRaLS.random_north_pole_point

CoRaLS.random_point_on_cap

CoRaLS.random_point_on_cicular_orbit

CoRaLS.random_point_on_elliptical_orbit

CoRaLS.random_point_on_elliptical_orbit

CoRaLS.random_point_on_sphere

CoRaLS.random_south_pole_point

CoRaLS.random_surface_normal

CoRaLS.random_surface_normal

CoRaLS.random_vector

CoRaLS.raytrace

CoRaLS.regolith_density

CoRaLS.regolith_density

CoRaLS.regolith_density

CoRaLS.regolith_field

CoRaLS.regolith_field

CoRaLS.regolith_index

CoRaLS.regolith_index

CoRaLS.regolith_index

CoRaLS.retrigger

CoRaLS.rician

CoRaLS.rician_trigger

CoRaLS.sample_auger

CoRaLS.sample_power_law

CoRaLS.sample_power_law

CoRaLS.save_acceptance

CoRaLS.sky_temperature

CoRaLS.snell_θt

CoRaLS.spherical_cap_area

CoRaLS.spherical_to_cartesian

CoRaLS.surface_transmission

CoRaLS.surface_transmission

CoRaLS.throw_cosmicray

CoRaLS.throw_cosmicray

CoRaLS.trials_passed

CoRaLS.trigger_all