acceptance.jl: The Main Monte Carlo Loop

The acceptance.jl is part of the CoRaLS module and contains functions and structures for calculating the acceptance of the system for sub-surface Ultra-High-Energy Cosmic Ray (UHECR) reflections from ice. It also includes utilities for analyzing trials and computing differential spectra.

Main Components

CoRaLS.Acceptance — Type

A structure to hold results of acceptance calculations, including trials, altitude, energies, and different acceptance types (direct and reflected).

Missing docstring.

Missing docstring for CoRaLS.acceptance. Check Documenter's build log for details.

CoRaLS.trials_passed — Function

trials_passed

Calculate the number of successful trials in the direct and reflected acceptances.

CoRaLS.differential_spectrum — Function

differential_spectrum(energies, AΩ, T)

Calculate the differential spectrum of detected UHECR events based on acceptance calculations and T, the total observation time of the experiment.

Calculates the differential spectrum of detected UHECR events .