Other Modules

Other CoRaLS modules not currently in use.

moon.jl

Models lunar geography, e.g. specific locations of permanently shadowed regions (PSRs).

CoRaLS.LunarPole — Type

Identify a lunar pole, North or South.

CoRaLS.latlon_to_xy — Method

latlon_to_xy(::LunarPole, lat, lon)

Convert latitude and longitude coordinates into (x, y) coordinates based on stereographic projections for either the North or South Pole.

Arguments

::LunarPole: Specify NorthPole or SouthPole.

lat: Latitude in degrees.

lon: Longitude in degrees.

Returns

(x, y): Coordinates in meters in the respective pole's stereographic projection.

Formulas from: http://pds.lroc.asu.edu/data/LRO-L-LROC-5-RDR-V1.0/LROLRC_2001/DOCUMENT/RDRSIS.PDF

CoRaLS.point_impacts_psr — Method

point_impacts_psr(surface)

Check if a point on the lunar surface impacts a permanently shadowed region (PSR).

Arguments

surface: A point on the lunar surface.

Returns

true if the point impacts a PSR, false otherwise.

CoRaLS.psr_area — Method

psr_area(altitude)

Calculate the total area of lunar permanently shadowed regions (PSRs) that are observable from a given altitude.

Arguments

altitude: Altitude of the observing instrument above the lunar surface.

Returns

Total area of PSRs visible from the given altitude.

CoRaLS.psr_fraction — Method

psr_fraction(altitude)

Determine the fraction of the observable lunar surface at a given altitude that is covered by PSRs.

Arguments

altitude: Altitude of the observing instrument above the lunar surface.

Returns

Fraction of the observable area covered by PSRs.

Raytrace

A raytracing version of the ray propagation model.

CoRaLS.raytrace — Method

raytrace(θ_i, Xmax, ice_depth, indexmodel)

Raytrace a ray from the surface with zenith angle (in the regolith)
of `θ_i` and calculate the zenith angle at the depth of `Xmax` and
at `ice_depth` using `indexmodel`.

Arguments

θ_i: Initial zenith angle of the ray at the lunar surface (in the regolith).

Xmax: Depth at which the cosmic ray reaches its maximum development (Xmax).

ice_depth: Depth of the ice layer beneath the lunar surface.

indexmodel: Model of the regolith index used for ray tracing.

Returns

Drego: Total path length of the ray within the regolith.

θXmax: Zenith angle of the ray at the Xmax depth.

θice: Zenith angle of the ray at the ice layer depth.

Description

This function performs ray tracing for a cosmic ray incident on the lunar surface, considering refraction effects as the ray passes through the regolith and approaches the ice layer. The function assumes a perfectly specular reflection at the ice layer, meaning the incident and reflection angles are equal. It computes the path length and zenith angles at crucial depths (Xmax and ice layer) and returns these values along with the total path length within the regolith.

Simulate_vector

Uses vector geometry and a mimimization routine to compute the launch angle that would cause the emergent ray to intersect the location of the spacecraft from a Direct or Reflected event.