simulate.jl: Main Cosmic Ray Simulation

The simulate.jl module simulates cosmic ray events and progates the RF signal through the regolith.

CoRaLS.AbstractSignal — Type

Abstract type for simulated signals, both direct and reflected.

CoRaLS.Direct — Type

The information stored for a direct RF detection.

CoRaLS.EventGeometry — Type

An enum for possible event geometries (i.e. direct or reflected)

CoRaLS.GeometryImplementation — Type

Abstract type for event simulation implementation.

CoRaLS.Reflected — Type

The information stored for a reflected RF detection.

CoRaLS.ScalarGeometry — Type

Approximate scalar geometry based on Andres' MC.

This makes the approximation that the CR impact point is the same as the refraction point - this is an error of order <10m out of the 30km distance to the payload.

CoRaLS.TrialFailed — Type

An enum for possible reasons why the a cosmic ray trial failed.

CoRaLS.VectorGeometry — Type

3D vector geometry implementation.

This is the original method of the UH MC - there are no approximations used in this solution.

Base.show — Method

Simplifiy the printing of Signal structs.

CoRaLS.compute_direct — Method

Compute the 'direct' RF solution using the scalar geometry.

CoRaLS.compute_reflected — Method

Compute the 'reflected' RF solution using the scalar solution.

CoRaLS.propagate_cosmicray — Method

Propagate cosmic ray into surface and compute direct and reflected RF.

CoRaLS.throw_cosmicray — Method

Simulate a single cosmic ray trial with a given energy.

This calculates the field at the payload for the direct and reflected emission, (if they exist) and returns whether the event triggered.

CoRaLS.throw_cosmicray — Method

Simulate a single cosmic ray trial with a given energy.

This calculates the field at the payload for the direct and reflected emission, (if they exist) and returns whether the event triggered.