Modules used in acceptance.jl

The main sub-modules called by acceptance.jl.

cosmic_ray.jl

Models cosmic rays and interactions. Computes depth of shower maximum (Xmax), and charge excess profile of cosmic ray-induced showers.

CoRaLS.charge_excess_profile — Method

charge_excess_profile(Ecr::Unitful.Quantity)

Return the profile of excess shower charge (in Coulombs) for a cosmic ray of energy Ecr.

This function calculates the charge profile based on parameters derived from ZHAireS simulations of cosmic ray air showers.

Arguments

Ecr: Energy of the cosmic ray in EeV.

Returns

Function Q(z) that evaluates the shower charge profile at a given distance z.

CoRaLS.estimate_Xmax — Method

estimate_Xmax(Ecr::Unitful.Quantity)

Estimate the mean and variation in Xmax for a given cosmic ray energy Ecr.

Xmax is the maximum depth in the atmosphere where a cosmic ray-induced air shower reaches its peak intensity. This function uses the Auger fits to Xmax and σ(Xmax) from: https://www.icrc2019.org/uploads/1/1/9/0/ 119067782/yushkovaugericrc2019mass1talk_final.pdf

This is only valid up to 1e20.5 (VERIFY THIS).

Arguments

Ecr: Energy of the cosmic ray in EeV.

Returns

Tuple of (mean, sigma) of Xmax in g/cm^2.

CoRaLS.propagate_to_Xmax — Method

propagate_to_Xmax(origin, direction, Ecr, densitymodel)

Propagate a cosmic ray from the surface to its shower maximum (Xmax).

This function calculates the point of shower maximum based on the cosmic ray's energy and direction, considering the density model of the medium through which it propagates.

Arguments

origin: Starting point of the cosmic ray.

direction: Direction vector of the cosmic ray.

Ecr: Energy of the cosmic ray.

densitymodel: Model of the density of the medium.

Returns

The position vector where the cosmic ray reaches Xmax.

efield.jl

Model and scale the electric field from a supplied cosmic ray event lookup table file.

CoRaLS.ARW — Type

The "ARW" model from Practical & Accurate Calculations...

CoRaLS.FORTE — Type

The FORTE (N. Lehtinen, PRD 2004) field model.

This requires an argument to determine what charge profile to use - currently Gaussian or GaisserHillas

CoRaLS.FieldModel — Type

This type is the base for all instances of electric field models for the cosmic ray emission.

CoRaLS.GaisserHillasProfile — Type

Gaisser-Hillas charge excess profile.

CoRaLS.GaussianProfile — Type

Gaussian charge excess profile.

CoRaLS.JAM — Type

The JAM field model.

This is model is wrong and should never be used for production results. The angular width is way too wide.

Parameterization from Alvarez-muniz et al arxiv:astro-ph/0512337 2005.

CoRaLS.ProfileModel — Type

This type is the base for all longitudinal profile models.

CoRaLS.ARW_RE — Function

The data file for the practical and accurate calculation.

CoRaLS.attenuation_length — Method

attenuation_length(ν, n, density; tanδnorm=6.5e-4)

Calculate the attenuation length for radio waves in regolith at given frequencies. This function considers the refractive index and density of the regolith, as well as the loss tangent, to estimate how far radio waves can travel before being significantly attenuated.

Arguments

ν: Array of frequencies.

n: Refractive index of the medium.

density: Density of the medium.

tanδnorm: Normalized loss tangent (default: 6.5e-4).

Returns

Array of attenuation lengths corresponding to each frequency.

CoRaLS.load_ARW — Method

load_ARW()

Load the ARW (Askaryan Radio Wave) electric field data and construct an interpolator for the electric field. This function is responsible for reading the ARW data from a file and setting up a linear interpolation model for later use in electric field calculations.

Returns

A function that takes an angle (θ) and frequency (ν) and returns the interpolated electric field value.

CoRaLS.regolith_field — Method

regolith_field(::ARW, Ecr, θ, Drego, Dvacuum; kwargs...)

Calculate the integrated electric field at the payload using the ARW model. This function takes into account the cosmic ray energy, viewing angle, and distances through regolith and vacuum to compute the electric field.

Arguments

Ecr: Energy of the cosmic ray.

θ: Viewing angle.

Drego: Distance through regolith.

Dvacuum: Distance through vacuum.

kwargs: Additional keyword arguments for fine-tuning the model.

Returns

Tuple of frequencies and corresponding electric field values.

CoRaLS.regolith_field — Method

regolith_field(::JAM, Ecr, θ,
                    Drego, Dvacuum;
                    n = sqrt(2.8), ν_min=30.0MHz, ν_max=300.0MHz, dν=10.0MHz,
                    Latten = 35.0m)

Calculate the integrated electric field at the payload using the JAM model.

Do not use this - this vastly overestimates the width of the "beam", and allows significantly more events to pass.

fresnel.jl

Computes fresnel coefficients and wave divergence at the surface.

CoRaLS.DivergenceModel — Type

Abstract type for a divergence model for surface refraction.

CoRaLS.FarFieldDivergence — Type

Far-field (spherical) divergence.

This is taken from the FORTE paper.

CoRaLS.MixedFieldDivergence — Type

Somewhere between near- and far- field divergence.

This is currently approximated by an additional sqrt(n) factor out front of the far-field (spherical) divergence.

CoRaLS.NearFieldDivergence — Type

Near-field (cylindrical) divergence.

This is currently approximated by an additional n factor out front of the far-field (spherical) divergence.

CoRaLS.divergence_tpar — Method

divergence_tpar(θ_i, n, Drego, Dvacuum)

Calculate modified Fresnel transmission coefficients for the parallel polarization, considering far-field (spherical) divergence of waves at the surface.

    # Arguments
    - `θ_i`: Incident angle.
    - `n`: Refractive index of the medium.
    - `Drego`: Distance through regolith.
    - `Dvacuum`: Distance through vacuum.

Note: We construct this so that it can be directly multiplied to an electric field, so have to back out the total distance, before doing the scaling.

Note: Since the refraction occurs in the near field (Fresnel region), the standard far-field (i.e. spherical) divergence assumption in the FORTE paper doesn't hold. Our current model for this is an additional factor of n (which boosts the transmission).

See the FORTE paper, Appendix 2, for more details. arXiV:0309656

CoRaLS.divergence_tperp — Method

divergence_tperp(::FarFieldDivergence, θ_i, n, Drego, Dvacuum)

Calculate modified Fresnel transmission coefficients for the perpendicular polarization, considering far-field (spherical) divergence of waves at the surface.

Arguments

θ_i: Incident angle.

n: Refractive index of the medium.

Drego: Distance through regolith.

Dvacuum: Distance through vacuum.

Returns

Modified Fresnel transmission coefficient for perpendicular polarization.

Note: We construct this so that it can be directly multiplied to an electric field, so have to back out the total distance, before doing the scaling.

Note: Since the refraction occurs in the near field (Fresnel region), the standard far-field (i.e. spherical) divergence assumption in the FORTE paper doesn't hold. Our current model for this is an additional factor of n (which boosts the transmission).

See the FORTE paper, Appendix 2, for more details. arXiV:0309656

CoRaLS.fresnel_coeffs — Method

fresnel_coeffs(θi, ni, nt; simple_t = true)

Calculate the parallel and perpendicular Fresnel reflection and transmission coefficients.

Each coefficient is the ratio of reflected/transmitted to incident amplitude (denoted r and t, not to be confused with fresnel coefficients for power, R and T)

If simplet is true (default) then the transmission coefficients are calculated as follows, otherwise they use the full Fresnel equations. tparallel = ni * (rparallel + 1) / nt tperpendicular = r_perpendicular + 1

Arguments

θi: Incident angle.

ni: Refractive index of the initial medium.

nt: Refractive index of the transmitting medium.

Returns

(rpar, rperp, tpar, tperp): Fresnel coefficients.

CoRaLS.fresnel_critical — Method

fresnel_critical(ni, nt)

Calculate the critical angle for total internal reflection.

Arguments

ni: Refractive index of the initial medium.

nt: Refractive index of the transmitting medium.

Returns

Critical angle for total internal reflection.

CoRaLS.snell_θt — Method

θt(θi, ni, nt)

Calculate the transmitted angle from Snell's Law.

geometry.jl

Model the geometry of the Moon and local

CoRaLS.cartesian_to_latlon — Method

Convert a cartesian point (x, y, z) into (lat, lon).

CoRaLS.cartesian_to_latlonalt — Method

Helper function to convert orbital position (x, y, z) to (lat, lon, altitude).

CoRaLS.cartesian_to_spherical — Method

Convert a cartesian point (x, y, z) into a spherical point (theta, phi, r).

CoRaLS.horizon_angle — Method

horizon_angle(altitude; R=Rmoon)

Calculate the angle below the horizontal of the horizon from a given altitude.

This returns "negative" angles, in radians, for "below" the horizontal.

Arguments

height: The altitude above the surface (with units).`

CoRaLS.intersect_with_sphere — Method

intersect_with_sphere(start, direction, radius)

Propagate a vector from start to radius along direction.

Assumes direction is normalized.

See the below link for the formalism behind this implementation: https://www.scratchapixel.com/lessons/3d-basic-rendering/ minimal-ray-tracer-rendering-simple-shapes/ray-sphere-intersection

CoRaLS.latlon_to_cartesian — Method

Convert a (lat, lon) to cartesian (x, y, z) vector at radius r.

CoRaLS.parse_orbit — Function

parse_orbit(fname)

Parse orbital info from a CSV file with columns: time, longitude, latitude, altitude.

Returns SampledOrbit matrix of lon,lat,alt positions to randomly sample.

CoRaLS.propagate_and_refract — Method

propagate_and_refract(start, antenna, direction, scale, indexmodel)

Propagate from start along direction to antenna, refracting at the surface.

CoRaLS.random_angles_on_cap — Method

Generate a (theta, phi) pair on a spherical cap uniformly in solid angle.

All angles are in radians.

CoRaLS.random_direction — Function

random_direction()

Generate a random Cartesian direction distributed uniformly on the unit-sphere with the pole of the sphere aligned with normal weighted by cos^2( heta)

CoRaLS.random_north_pole_point — Method

random_north_pole_point()

Draw a random point on the North lunar pole

CoRaLS.random_point_on_cap — Method

Sample a random point on a spherical cap uniformly in solid angle.

We assume that the spherical cap is centered on the +z-pole of the sphere.

CoRaLS.random_point_on_cicular_orbit — Function

Return cartesian vector to spacecraft at altitude.

Note: this samples uniformly in latitude which does not give uniformly distributed points on the sphere (overdensity of points on the poles which is expected for circular polar orbits).

CoRaLS.random_point_on_elliptical_orbit — Function

Return cartesian vector to spacecraft at altitude.

Note: this samples uniformly in latitude which does not give uniformly distributed points on the sphere (overdensity of points on the poles which is expected for circular polar orbits).

CoRaLS.random_point_on_elliptical_orbit — Method

random_point_on_elliptical_orbit(periapse, apoapse, inclination; rbody=Rmoon)

Return cartesian vector to spacecraft on an elliptical orbit.

Parameters:

periapse: minimum altitude above reference body surface

apoapse: maximum altitude above reference body surface

inclination: orbital inclination in degrees

rbody: radius of the central body (default: Rmoon)

Note: This samples uniformly in true anomaly which gives a non-uniform distribution of points along the orbit (points are more densely distributed near periapse than apoapse, which is physically correct for constant angular momentum).

CoRaLS.random_point_on_sphere — Function

random_point_on_sphere()

Return cartesian point (x, y, z) randomly distributed on sphere of radius r.

This uses the triple gaussian method which is slightly faster than the spherical coords method. https://mathworld.wolfram.com/SpherePointPicking.html

CoRaLS.random_south_pole_point — Method

random_south_pole_point()

Draw a random point on the South lunar pole

CoRaLS.random_vector — Function

random_vector()

Generate a random Cartesian direction distributed uniformly on the unit-sphere with the pole of the sphere aligned with normal

CoRaLS.rotate_vector_to_align — Method

rotate_vector_to_align(vec, from, to)

Rotate vec so that the direction from is aligned with to. If from and to are already aligned, returns vec unchanged. Numerically stable for nearly parallel vectors.

CoRaLS.spherical_cap_area — Function

spherical_cap_area(theta, r)

Area of a spherical cap with central angle theta and radius r.

CoRaLS.spherical_to_cartesian — Method

Convert a spherical point (theta, phi, r) into Cartesian coordinates.

ice.jl

Models interactions of the RF signal with a variably rough ice layer.

CoRaLS.GaussianIceRoughness — Type

GaussianIceRoughness(σ::typeof(0.0cm)) <: IceRoughnessModel

A model representing Gaussian ice surface roughness, characterized by a standard deviation (σ) of the angular distribution of the surface.

Arguments

σ: Standard deviation of the surface angles, representing the degree of roughness.

CoRaLS.IceRoughnessModel — Type

ice.jl

Models the surface roughness of ice, which is a critical factor in simulating the reflection and transmission of radio waves on ice surfaces. It includes abstract and concrete types representing different models of ice roughness and functions to apply these models to simulated electric fields.

Main Components

IceRoughnessModel: An abstract type for general ice roughness models.

NoIceRoughness: A model representing no surface roughness (smooth ice).

GaussianIceRoughness: A model representing Gaussian ice surface roughness, characterized by a standard deviation of surface angles.

ice_roughness: Functions to apply roughness models to electric fields, modifying them based on the specified roughness model.

CoRaLS.NoIceRoughness — Type

NoIceRoughness <: IceRoughnessModel

Represents a model of ice surface with no roughness (smooth surface). This model assumes a perfectly smooth ice surface, affecting the reflection and transmission of radio waves accordingly.

CoRaLS.ice_roughness — Method

ice_roughness(model::IceRoughnessModel, ν, E, θ_i)

Apply the specified ice roughness model to a simulated electric field. The function modifies the electric field based on the roughness characteristics of the model.

Arguments

model: An instance of IceRoughnessModel to apply.

ν: Frequencies of the electric field.

E: Electric field amplitudes.

θ_i: Angle of incidence at the ice surface.

Returns

Modified frequencies and electric field based on the roughness model. If the model is NoIceRoughness, the function returns the given frequencies and electric field. If the model is GaussianIceRoughness, the function modifies the electric field based on the "Diffuse reflection by rough surfaces: an introduction" by Sylvain, Pg. 671.

regolith.jl

Models the lunar regolith and propagation of RF signal through it.

CoRaLS.CE3Index — Type

PL NOTE: Chang'E-3 regolith index

CoRaLS.CE4Index — Type

PL NOTE: Chang'E-4 regolith index

CoRaLS.ConstantDensity — Type

Constant regolith density

CoRaLS.ConstantIndex — Type

Constant regolith refractive index.

CoRaLS.DivinerIndex — Type

PL NOTE: Diviner regolith index

CoRaLS.LunarSourceBookIndex — Type

PL NOTE: Lunar Source Book regolith index

CoRaLS.OldIncorrectDensity — Type

A logarithmic regolith density model from PG's ARIA MC.

This is WRONG - do NOT use this except when trying to reproduce or evaluate results from early August 2021.

CoRaLS.RegolithDensity — Type

Abstract type for regolith density models.

CoRaLS.RegolithIndex — Type

Abstract type for regolith index models.

CoRaLS.StrangwayDensity — Type

A regolith density model from Olhoeft & Strangway.

This is "Curve A" from O&S.

CoRaLS.StrangwayDensityCB — Type

A regolith density model from Olhoeft & Strangway.

This is "Curve B" from O&S.

CoRaLS.StrangwayIndex — Type

A refractive index model from Olhoeft & Strangway.

This uses Eq. 1 from Olhoeft and Strangway.

CoRaLS.StrangwayIndexCB — Type

PL NOTE: Olhoeft and Strangway Curve B regolith index

CoRaLS.SurfaceDeepIndex — Type

A "surface" and "deep" regolith model.

This uses a small refractive index at the surface and a larger one for below the surface.

CoRaLS.create_density_lut — Function

create_density_lut(N)

Create a LUT of the density as a function of depth using "Curve B" from Olhoeft & Strangway. N is the number of points in the LUT.

CoRaLS.create_density_lut — Function

create_density_lut(N)

Create a LUT of the density as a function of depth using "Curve A" from Olhoeft & Strangway. N is the number of points in the LUT.

CoRaLS.regolith_density — Method

regolith_density(::CE3LPRDensity, depth)

Calculate the density of the regolith in g/cm^3 for depth.

CoRaLS.regolith_density — Method

regolith_density(::OldIncorrectDensity, depth)

Calculate the density of the regolith in g/cm^3 for depth.

CoRaLS.regolith_density — Method

regolith_density(::ConstantDensity, depth)

Calculate the density of the regolith in g/cm^3.

CoRaLS.regolith_density — Method

regolith_density(::StrangwayDensity, depth)

Calculate the density of the regolith in g/cm^3 for depth using the Olhoeft & Strangway curve.

CoRaLS.regolith_density — Method

regolith_density(::StrangwayDensity, depth)

Calculate the density of the regolith in g/cm^3 for depth using the Olhoeft & Strangway curve.

CoRaLS.regolith_index — Function

regolith_index(::IndexModel, depth, low_temp_corr_factor=0.9)

A refractive index model from Olhoeft & Strangway.

Dielectric constant fit is done by Olhoeft and Strangway Peter estimated a 10% reduction for lunar PSR's due to the ~80 K temperatures compared to the typical lunar temperatures of O&S, hence the lowtempcorr_factor.

CoRaLS.regolith_index — Method

regolith_index(::ConstantIndex, depth)

A constant refractive index as a function of depth.

CoRaLS.regolith_index — Method

regolith_index(::SurfaceDeepIndex, depth)

A two-part "surface" and "deep" refractive index model.

spectrum.jl

Cosmic ray spectrum (AUGER) calculations.

CoRaLS.auger_spectrum — Method

Randomly evaluate the Auger UHECR spectrum in eV^-1 km^-2 sr^-1 yr^-1 within experimental uncertainties.

This method differs from the one argument version of auger_spectrum() in that it randomly samples the parameters within experimental uncertainties so that you can capture the variation in the cosmic ray spectrum.

This uses the parameterization from: https://academic.oup.com/ptep/article/2017/12/12A103/4665686

CoRaLS.auger_spectrum — Method

Evaluate the Auger UHECR spectrum in eV^-1 km^-2 sr^-1 yr^-1.

This uses the parameterization from: https://academic.oup.com/ptep/article/2017/12/12A103/4665686

CoRaLS.auger_spectrum_2020 — Method

Evaluate the Auger UHECR spectrum in eV^-1 km^-2 sr^-1 yr^-1.

Valid for 10^18 to ~10^20 eV.

See Equation 9 in: Measurement of the cosmic-ray energy spectrum above 2.5e18 eV using the Pierre Auger Observatory. Phys.Rev.D (2020) https://doi.org/10.1103/PhysRevD.102.062005

CoRaLS.auger_spectrum_2021 — Method

Evaluate the Auger UHECR spectrum in eV^-1 km^-2 sr^-1 yr^-1.

Valid for 10^17 to ~10^20 eV.

See Equation 9 in: Abreu, P., Aglietta, M., Albury, J.M. et al. The energy spectrum of cosmic rays beyond the turn-down around 10^17 eV as measured with the surface detector of the Pierre Auger Observatory. Eur. Phys. J. C 81, 966 (2021). https://doi.org/10.1140/epjc/s10052-021-09700-w

CoRaLS.sample_auger — Method

sample_auger(min_energy, max_energy)

Draw a random sample from the Auger UHECR flux between min_energy and max_energy.

CoRaLS.sample_power_law — Method

sample_power_law(gamma, N; min_value, max_value)

Draw N random samples from a power law with exponent gammma.

If provided, min_value and max_value set the range over which the power-law is sampled (defaults to 1 EeV and 1000 EeV)

This uses the prescription from: https://mathworld.wolfram.com/RandomNumber.html to convert a uniform deviate into a power law with a given exponent.

CoRaLS.sample_power_law — Method

sample_power_law(gamma; min_value, max_value)

Draw random samples from a power law with exponent gammma.

If provided, min_value and max_value set the range over which the power-law is sampled (defaults to 1 EeV and 1000 EeV)

This uses the prescription from: https://mathworld.wolfram.com/RandomNumber.html to convert a uniform deviate into a power law with a given exponent.

surface.jl

Models surface roughness and slope, including generating random surface normals and calculating surface transmission.

CoRaLS.GaussianRoughness — Type

Gaussian surface roughness model.

This is characterized by Gaussian "sigma" of the angular distribution of the surface.

CoRaLS.GaussianSlope — Type

Gaussian surface slope model.

This is characterized by Gaussian "sigma" of the angular distribution of the surface in degrees.

CoRaLS.NoRoughness — Type

No surface roughness (smooth)

CoRaLS.NoSlope — Type

No surface slope (smooth)

CoRaLS.RoughnessModel — Type

Abstract type for regolith surface roughness model.

CoRaLS.SlopeModel — Type

Abstract type for regolith surface slope model.

CoRaLS.random_surface_normal — Method

random_surface(slope::GaussianSlope, normal)

Generate a random normal vector for a surface slope that would originally be pointing along normal in the absence of roughness.

CoRaLS.random_surface_normal — Method

random_surface(slope::NoSlope, normal)

Generate a random normal vector for a surface slope that would originally be pointing along normal in the absence of roughness.

This just returns normal since this implements a "No Slope" model.

CoRaLS.surface_transmission — Method

surface_transmission(::GaussianRoughness, ν, E, θ_i)

Apply a simple model for surface transmission through a rough surface by taking the average of the transmission coefficient over the Gaussian

This is based on the "Diffuse reflection by rough surfaces: an introduction" by Sylvain, Pg. 671

CoRaLS.surface_transmission — Method

surface_transmission(::NoRoughness, ν, E, θ_i)

Apply roughness to a simulated electric field given the frequencies, the electric field, and the angle of incidence at the surface. This returns the frequencies and electric field back to the caller.

Without roughness, we just return ν and E as given.

trigger.jl

Simple triggers for cosmic ray detection.

CoRaLS.gaussian_trigger — Function

gaussian_trigger(peak=67μV/m, σ=45.0; dν=10.0MHz, θ0=-90.)

Creates a Gaussian beam threshold pointing down towards the nadir with peak threshold at the nadir and a width of σ in degrees. This uses the current estimate for a sinuous antenna array by P. Gorham

Arguments

peak=67μV/m: Peak threshold of the electric field at the nadir.

σ=45.0: Width of the Gaussian beam in degrees.

dν=10.0MHz: Frequency step size.

θ0=-90.: Nadir angle in degrees.

Returns

A function that checks if an event's electric field, after being scaled by a Gaussian factor based on its off-axis angle, exceeds the peak threshold.

Description

This trigger function incorporates a Gaussian beam pattern, scaling the threshold for the electric field magnitude based on the event's off-axis angle relative to the nadir. The Gaussian factor decreases the effective threshold as the angle increases from the nadir.

Usage

Useful in scenarios where the sensitivity of detection varies with the angle of incident cosmic rays, simulating a realistic antenna response.

CoRaLS.magnitude_trigger — Function

magnitude_trigger(peak=67μV/m; dν=10.0MHz)

Create a trigger function that checks if the total magnitude of the electric field is greater than some threshold.

Arguments

peak=67μV/m: The threshold for the electric field's magnitude.

dν=10.0MHz: Frequency step size.

Returns

A function that checks if the norm of the electric field of an event is greater than the threshold.

Description

This function generates a trigger function that evaluates whether the total magnitude of the electric field for a cosmic ray event exceeds a specified threshold. The electric field magnitude is calculated as the product of the field's polarization norm, frequency step size, and sum of electric field components.

Usage

Used to filter events based on the strength of their electric field magnitude, allowing selection of significant cosmic ray events.

CoRaLS.rician_trigger — Method

rician_trigger(;snr_threshold=5.0, bw=900.0MHz, peak=calculated, σ=45.0)

Create a Rician beam trigger function based on signal-to-noise ratio.

Parameters

snr_threshold=5.0: The signal-to-noise ratio threshold for triggering.

bw=900.0MHz: Bandwidth for the trigger.

peak=calculated: Peak threshold calculated based on the provided parameters.

σ=45.0: Width of the Gaussian beam in degrees.

Returns

A trigger function that evaluates if the signal-to-noise ratio of an event's electric field exceeds the specified threshold.

Description

This trigger function calculates a scaled threshold based on the event's off-axis angle and evaluates if the event's electric field, subjected to a Rician distributed noise model, surpasses this threshold. The function accounts for both primary and subsurface reflections.

Usage

Ideal for more sophisticated detection scenarios where the signal-to-noise ratio plays a crucial role in determining event significance.

CoRaLS.trigger_all — Method

trigger_all()

Always triggers. Used for testing geometric acceptance and debugging.

detector.jl

Detector models for complex triggering with antennas and SNR levels.

CoRaLS.ANITA — Method

ANITA(;kwargs...)

Create a model for the ANtarctic Impulsive Transient Antenna (ANITA) used in cosmic ray detection simulations. Configures the ANITA system based on given parameters or defaults.

Keyword Arguments

kwargs: Parameters for the ANITA model, including frequency range and other antenna-specific settings.

Returns

A configured ANITA antenna model.

CoRaLS.LPDA — Method

LPDA(;kwargs...)

Create a model for the Low-Profile Dipole Array (LPDA) used in cosmic ray detection simulations. Configures the LPDA based on given parameters or defaults.

Keyword Arguments

kwargs: Parameters for the LPDA model, such as altitude, frequency range, and noise settings.

Returns

A configured LPDA antenna model.

CoRaLS.create_antenna — Method

create_antenna(fname, skyfrac; kwargs...)

Construct an antenna-factor based antenna simulation. This function sets up the parameters for an antenna system, including its frequency range, noise characteristics, and response factors, based on antenna factor data from a file.

Arguments

fname: Filename containing antenna factor data.

skyfrac: Fraction of the antenna's view that is sky.

kwargs: Additional keyword arguments including altitude, frequency range, angular resolutions, and trigger settings.

Returns

A function that evaluates the antenna trigger condition for a given event.

Sky

Computes the sky temperature for noise calculations in detector.jl.

CoRaLS.extragalactic_noise — Method

extragalactic_noise(ν)

Calculate the extragalactic contribution to RF background noise (in W m^-2 Hz^-1 sr^-1) as function of frequency (in MHz) near the galactic poles.

Arguments

ν: Frequency at which the extragalactic noise is to be calculated.

Returns

The extragalactic noise in W m^-2 Hz^-1 sr^-1.

Description

This function estimates the extragalactic noise contribution based on the frequency. It uses a parameterization taken from Dulk (2001), specifically tailored for observations near the galactic poles: https://www.aanda.org/articles/aa/full/2001/02/aads1858/aads1858.right.html

CoRaLS.galactic_noise — Method

galactic_noise(ν)

Calculate the galactic contribution to RF background noise as function of frequency near the galactic poles.

Arguments

ν: Frequency at which the galactic noise is to be calculated.

Returns

The galactic noise in W m^-2 Hz^-1 sr^-1.

Description

This function computes the galactic noise component of the RF background. The calculation is based on a model from Dulk (2001): https://www.aanda.org/articles/aa/full/2001/02/aads1858/aads1858.right.html

CoRaLS.sky_temperature — Method

sky_temperature(ν)

Calculate the brightness temperature near the galactic poles as a function of frequency.

Arguments

ν: Frequency at which the sky temperature is to be calculated.

Returns

The brightness temperature near the galactic poles in Kelvin.

Description

This function calculates the combined extragalactic and galactic noise contributions to the radio frequency background and converts it into a brightness temperature. This temperature is a crucial parameter in astrophysical studies, especially when analyzing cosmic ray interactions and detections near the galactic poles.

utils.jl

Miscellaneous utilities.

CoRaLS.polarization_angle — Method

polarization_angle(events)

Calculate the polarization angle for a given set of events.

Arguments

events: A collection of event data, each with a polarization vector.

Returns

An array of polarization angles in degrees.

Description

This function computes the polarization angle for each event in the provided collection. The polarization angle is calculated relative to the horizontal plane of the antenna array, with 0 degrees corresponding to horizontal polarization and 90 degrees to vertical polarization.

CoRaLS.retrigger — Method

retrigger(AΩ, trig)

Recalculate the effective acceptance for direct and reflected events.

Arguments

AΩ: A struct holding initial acceptance calculation results.

trig: A trigger function used to reassess event acceptance.

Returns

A tuple containing the recalculated effective acceptances for direct and reflected events.

Description

This function applies a new trigger to a set of cosmic ray events, recalculating the effective acceptance based on which events would pass this new trigger. It's useful for assessing the impact of different trigger criteria on event detection rates.

CoRaLS.rician — Function

rician(ν, σ, N = 1)

Draw one or more Rician random variables.

Arguments:

ν: The noncentrality parameter, can be a scalar or an array.

σ: The scale parameter, can be a scalar or an array matching the size of ν.

N = 1: The number of samples to draw. Used only if ν is a scalar.

Returns

If ν and σ are single values, it returns a single Rician random variable.

If ν and σ are arrays, it returns an array of Rician random variables with each element corresponding to elements in ν and σ.

If ν is a single value and N > 1, it returns an array of N Rician random variables, all generated with the same ν and σ.

plots.jl

Visualization Tools for CoRaLS Simulation Data

CoRaLS.mcse — Method

mcse(count, ntrials, [gAΩ])

Compute the Monte Carlo Standard Error as count or acceptance if gAΩ is given.

Uses Monte Carlo Standard Error for binomial proportion: sqrt(p*(1-p)/n)

CoRaLS.plot_acceptance — Method

plot_acceptance(AΩ::Acceptance)

Plot the direct and reflected acceptance of the CoRaLS project.

Arguments:

AΩ::Acceptance: An object containing the acceptance calculation results.

This function generates a line plot illustrating the acceptance rates for both direct and reflected events across different energy levels.

CoRaLS.plot_differential_spectrum — Method

plot_differential_spectrum(AΩ::Acceptance, T::Unitful.Time)

Create a plot of the differential event spectrum of CoRaLS events given an acceptance object and a mission duration.

Arguments:

AΩ::Acceptance: An object containing the acceptance calculation results.

T::Unitful.Time: The total observation time of the experiment.

This function plots the estimated number of events per energy bin over the duration of the mission.

CoRaLS.plot_incident_angles — Method

plot_incident_angles(AΩ::Acceptance)

Plot the incident angles at the ice and the surface for both direct and reflected events, along with the corresponding Fresnel coefficients.

Arguments:

AΩ::Acceptance: An object containing the acceptance calculation results.

Generates a two-part plot. The left plot shows histograms of incident angles at the surface for direct and reflected events. The right plot shows incident angles at the ice layer for reflected events. Behind these histograms, the function plots the transmission and reflection coefficients as a function of the incident angle.

CoRaLS.plot_offaxis_angle — Method

plot_offaxis_angle(AΩ::Acceptance)

Generate a plot showing the off-axis angle distribution of direct and reflected events in the CoRaLS simulations.

Arguments:

AΩ::Acceptance: An object containing the acceptance calculation results.

CoRaLS.plot_polarization_angle — Method

plot_polarization_angle(AΩ::Acceptance)

Plot histograms of polarization angles for direct and reflected events.

Arguments:

AΩ::Acceptance: An object containing the acceptance calculation results.

CoRaLS.plot_rate_experiment — Method

plot_rate_experiment(AΩ::Acceptance, xs; xlabel)

Plot the counts in events / yr for an array of acceptance runs computed at some variable x (e.g., ice depths, altitudes).

Arguments:

AΩs : Array of acceptance objects.

xs : Array of paramter values for the x axis corresponding to each AΩ.

This function plots the estimated number of events per energy bin over the duration of the mission.