CoRaLS Documentation

Welcome to the CoRaLS documentation

CoRaLS — Module

CoRaLS

The CoRaLS.jl (Cosmic Ray Lunar Sounder) Monte Carlo model computes detection rates of Askaryan emission from cosmic ray interactions in lunar regolith.

Installation

Install Julia

Install Python and the matplotlib package, the easiest way is with Anaconda (e.g. conda list matplotlib).

Fork or clone this repository:

git clone git@github.com:cjtu/CoRaLS.jl.git

(First time only): Setup path to python/matplotlib. In the Julia REPL, supply the path to your python environment from #2 in quotes to ENV["PYTHON"]="" (leave blank to use system default python). Then add and build the PyCall package:

$ julia

julia> ENV["PYTHON"]=""
julia> using Pkg
julia> Pkg.add("PyCall")
julia> Pkg.build("PyCall")

Run the test suite to test the installation:

julia --project=/path/to/CoRaLS.jl /path/to/CoRaLS.jl/test/runtests.jl

If the tests were successful, CoRaLS.jl is compiled and ready to use!

Calculating rates with CoRaLS

Start Julia with the CoRaLS project active using the --project flag pointing to the CoRaLS directory (if you are in the directory use --project=.). For multithreaded mode, use -t to specify the number of threads (default is 1, "auto" chooses for you).

julia --project=/path/to/CoRaLS.jl -t "auto"
julia> using CoRaLS

import CoRaLS with:

julia> using CoRaLS

Compute and plot an acceptance:

julia> A = acceptance(10000, 20; region=create_region("psr:south"), spacecraft=CircularOrbit(50.0km))
julia> plot_acceptance(A)

See full documentation online at... (coming soon)

Developers

To make and preview docs locally run the following from the root CoRaLS directory:

julia --project=docs -e 'include("docs/make.jl"); using LiveServer; serve(dir="docs/build")'

To run tests:

julia --project=. test/runtests.jl

Citing

Paper coming soon.

API Index

CoRaLS

CoRaLS.ARW

CoRaLS.AbstractSignal

CoRaLS.Acceptance

CoRaLS.CE3Index

CoRaLS.CE4Index

CoRaLS.ConstantDensity

CoRaLS.ConstantIndex

CoRaLS.Direct

CoRaLS.DivergenceModel

CoRaLS.DivinerIndex

CoRaLS.EventGeometry

CoRaLS.FORTE

CoRaLS.FarFieldDivergence

CoRaLS.FieldModel

CoRaLS.GaisserHillasProfile

CoRaLS.GaussianIceRoughness

CoRaLS.GaussianProfile

CoRaLS.GaussianRoughness

CoRaLS.GaussianSlope

CoRaLS.GeometryImplementation

CoRaLS.IceRoughnessModel

CoRaLS.JAM

CoRaLS.LunarSourceBookIndex

CoRaLS.MixedFieldDivergence

CoRaLS.NearFieldDivergence

CoRaLS.NoIceRoughness

CoRaLS.NoRoughness

CoRaLS.NoSlope

CoRaLS.OldIncorrectDensity

CoRaLS.ProfileModel

CoRaLS.Reflected

CoRaLS.RegolithDensity

CoRaLS.RegolithIndex

CoRaLS.RoughnessModel

CoRaLS.ScalarGeometry

CoRaLS.SlopeModel

CoRaLS.StrangwayDensity

CoRaLS.StrangwayDensityCB

CoRaLS.StrangwayIndex

CoRaLS.StrangwayIndexCB

CoRaLS.SurfaceDeepIndex

CoRaLS.TrialFailed

CoRaLS.VectorGeometry

Base.show

CoRaLS.ANITA

CoRaLS.ARW_RE

CoRaLS.LPDA

CoRaLS.acceptance

CoRaLS.attenuation_length

CoRaLS.auger_spectrum

CoRaLS.auger_spectrum

CoRaLS.auger_spectrum_2020

CoRaLS.auger_spectrum_2021

CoRaLS.cartesian_to_latlon

CoRaLS.cartesian_to_latlonalt

CoRaLS.cartesian_to_spherical

CoRaLS.charge_excess_profile

CoRaLS.compute_direct

CoRaLS.compute_reflected

CoRaLS.create_antenna

CoRaLS.create_density_lut

CoRaLS.create_density_lut

CoRaLS.differential_spectrum

CoRaLS.divergence_tpar

CoRaLS.divergence_tperp

CoRaLS.estimate_Xmax

CoRaLS.extragalactic_noise

CoRaLS.fresnel_coeffs

CoRaLS.fresnel_critical

CoRaLS.galactic_noise

CoRaLS.gaussian_trigger

CoRaLS.horizon_angle

CoRaLS.ice_roughness

CoRaLS.intersect_with_sphere

CoRaLS.latlon_to_cartesian

CoRaLS.load_ARW

CoRaLS.load_acceptance

CoRaLS.magnitude_trigger

CoRaLS.mcse

CoRaLS.merge_acceptance

CoRaLS.old_acceptance

CoRaLS.parse_orbit

CoRaLS.plot_acceptance

CoRaLS.plot_differential_spectrum

CoRaLS.plot_incident_angles

CoRaLS.plot_offaxis_angle

CoRaLS.plot_polarization_angle

CoRaLS.plot_rate_experiment

CoRaLS.polarization_angle

CoRaLS.propagate_and_refract

CoRaLS.propagate_cosmicray

CoRaLS.propagate_to_Xmax

CoRaLS.random_angles_on_cap

CoRaLS.random_direction

CoRaLS.random_north_pole_point

CoRaLS.random_point_on_cap

CoRaLS.random_point_on_cicular_orbit

CoRaLS.random_point_on_elliptical_orbit

CoRaLS.random_point_on_elliptical_orbit

CoRaLS.random_point_on_sphere

CoRaLS.random_south_pole_point

CoRaLS.random_surface_normal

CoRaLS.random_surface_normal

CoRaLS.random_vector

CoRaLS.regolith_density

CoRaLS.regolith_density

CoRaLS.regolith_density

CoRaLS.regolith_density

CoRaLS.regolith_density

CoRaLS.regolith_field

CoRaLS.regolith_field

CoRaLS.regolith_index

CoRaLS.regolith_index

CoRaLS.regolith_index

CoRaLS.retrigger

CoRaLS.rician

CoRaLS.rician_trigger

CoRaLS.rotate_vector_to_align

CoRaLS.sample_auger

CoRaLS.sample_power_law

CoRaLS.sample_power_law

CoRaLS.save_acceptance

CoRaLS.sky_temperature

CoRaLS.snell_θt

CoRaLS.spherical_cap_area

CoRaLS.spherical_to_cartesian

CoRaLS.surface_transmission

CoRaLS.surface_transmission

CoRaLS.throw_cosmicray

CoRaLS.throw_cosmicray

CoRaLS.trials_passed

CoRaLS.trigger_all