acceptance.jl: The Main Monte Carlo Loop

The main acceptance() loop computes many cosmic ray impacts to a supplied region of the lunar surface and observed with a chosen spacecraft and triggering method, returning an Acceptance object containing the number of detected events and final acceptance in [$km^2$ sr] units.

CoRaLS.Acceptance — Type

A structure to hold results of acceptance calculations, including trials, altitude, energies, and different acceptance types (direct and reflected).

CoRaLS.acceptance — Method

acceptance(ntrials, nbins; region, spacecraft, trigger, ...)

Calculates the acceptance of CoRaLS for sub-surface UHECR reflections. It involves simulations of cosmic ray interactions, triggering conditions, and aggregating results across multiple trials and energy bins.

CoRaLS.differential_spectrum — Method

differential_spectrum(energies, AΩ, T)

Calculate the differential spectrum of detected UHECR events based on acceptance calculations and T, the total observation time of the experiment.

Calculates the differential spectrum of detected UHECR events .

CoRaLS.load_acceptance — Method

load_acceptance(filename::String)::Acceptance

Load the acceptance data from a .jld2 file.

CoRaLS.merge_acceptance — Method

merge_acceptance(A::Acceptance, B::Acceptance)::Acceptance

Merge two Acceptance structures into one, combining their trial counts and results.

CoRaLS.old_acceptance — Method

old_acceptance(ntrials, nbins; ...)

Calculates the acceptance of CoRaLS for sub-surface UHECR reflections. It involves simulations of cosmic ray interactions, triggering conditions, and aggregating results across multiple trials and energy bins.

CoRaLS.save_acceptance — Method

save_acceptance(A::Acceptance, filename::String)

Save the acceptance data to a file (only .jld2 supported).

CoRaLS.trials_passed — Method

trials_passed

Calculate the number of successful trials in the direct and reflected acceptances.